Rename hexagon_vtcm* to hexagon_vtcm_pool*
